API/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
gunicorn==21.2.0
